# static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "Your Flask API",
    "description": "API documentation for Your Flask API"
  },
  "basePath": "/",
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
    "/login": {
      "post": {
        "summary": "User Login",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              },
              "required": ["username", "password"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns a token for successful login",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                },
                "Token": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/houses": {
      "get": {
        "summary": "Get all houses",
        "responses": {
          "200": {
            "description": "Returns a list of houses",
            "schema": {
              "type": "object",
              "properties": {
                "houses": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/houseswithquery": {
      "get": {
        "summary": "Get houses with query parameters"
      }
    }
  }
}